fix: Slice of Rules type
The type of `r.Rules` has been changed from pointer type to value type
since SDK V2.

When a value is retrieved from value type using `for ... range`, the
pointer of the loop variable that stores the value is fixed.

That causes the problem that all elements of slice eventually refer to
the same pointer.
To avoid that problem, change the base of slice from pointer type to
value type.